+++ cmake/ADIOSBisonFlexSub.cmake
@@ -1,20 +1,6 @@
 FUNCTION (SETUP_ADIOS_BISON_FLEX_SUB)
 
-IF ((${CMAKE_SYSTEM_NAME} STREQUAL "Darwin") OR
-   (${CMAKE_SYSTEM_NAME} STREQUAL "Linux"))
-   set (BISON_FLEX_PRECOMPILE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/toolkit/derived/parser/pregen-source/Linux")
-elseif (${CMAKE_SYSTEM_NAME} STREQUAL "Windows")
-   set (BISON_FLEX_PRECOMPILE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/toolkit/derived/parser/pregen-source/Windows")
-else()
-   if (NOT BISON_FOUND)
-      message (FATAL_ERROR "Bison was not found and no pregenerated Bison/Flex"
-              "source is available for ${CMAKE_SYSTEM_NAME}. Please install Bison or Yacc")
-
-   else()
-      message (FATAL_ERROR "Flex was not found and no pregenerated Bison/Flex" 
-      	      "source is available for ${CMAKE_SYSTEM_NAME}. Please install Bison or Yacc")
-   endif()
-ENDIF()
+set (BISON_FLEX_PRECOMPILE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/toolkit/derived/parser/pregen-source")
 
 ADD_CUSTOM_COMMAND(OUTPUT parser.cpp
         COMMAND ${CMAKE_COMMAND} -E copy ${BISON_FLEX_PRECOMPILE_DIR}/parser.cpp ${CMAKE_CURRENT_BINARY_DIR}
